+++ src/dpr/middleware/setUpDprResources.ts
@@ -79,10 +79,15 @@ const setFeatureFlags = async (res: Response, featureFlagService: FeatureFlagSer
   const subject = getFeatureFlagEvaluationSubject(res)
 
   const currentFlags = res.app.locals['featureFlags'].flags
-  logger.info(`CURRENT_FEATURE_FLAGS: ${currentFlags}`)
+  logger.info('SAVE_DEFAULTS_DEBUG currentFlags', JSON.stringify(currentFlags))
+
   const newFlags = await featureFlagService.evaluateBooleanFlags(FEATURE_FLAGS, subject)
 
+  logger.info('SAVE_DEFAULTS_DEBUG newFlags', JSON.stringify(newFlags))
+
   const hasChanged = JSON.stringify(currentFlags) !== JSON.stringify(newFlags)
+
+  logger.info('SAVE_DEFAULTS_DEBUG hasChanged', hasChanged)
 
   if (hasChanged) {
     res.app.locals['featureFlags'].flags = newFlags
@@ -296,7 +301,12 @@ export function recordDefinitionsCheck(session: { lastDefinitionsCheck?: number 
  */
 const initialiseServices = async (services: Services, res: Response) => {
   const { dprUser } = localsHelper.getValues(res)
+
+  logger.info('SAVE_DEFAULTS_DEBUG', JSON.stringify({ dprUser }))
+
   if (dprUser.id) {
+    logger.info('SAVE_DEFAULTS_DEBUG initialising services')
+
     // Downloading
     if (!res.app.locals['downloadingEnabled']) {
       res.app.locals['downloadingEnabled'] = services.downloadPermissionService.enabled
@@ -339,18 +349,26 @@ const initialiseServices = async (services: Services, res: Response) => {
       logger.info(`Init service: reportIdMigrationService: ${res.app.locals['reportIdMigrationServiceEnabled']}`)
     }
 
-    // Save defaults
-    const enabled = isBooleanFlagEnabledOrMissing('saveDefaultsEnabled', res.app)
-      ? services.defaultFilterValuesService.enabled
-      : false
+    const serviceEnabled = services.defaultFilterValuesService.enabled
+    const flagEnabled = isBooleanFlagEnabledOrMissing('saveDefaultsEnabled', res.app)
+
+    logger.info(`SAVE_DEFAULTS_DEBUG`, JSON.stringify({ flagEnabled, serviceEnabled }))
+
+    const enabled = flagEnabled ? serviceEnabled : false
+
+    logger.info(`SAVE_DEFAULTS_DEBUG final enabled: ${enabled}`)
 
     const current = res.app.locals['saveDefaultsEnabled']
-    logger.info(`SAVE_DEFAULTS_DEBUG set: ${current}`)
+
+    logger.info(`SAVE_DEFAULTS_DEBUG local saveDefaultsEnabled: ${current}`)
+
     if (current !== enabled) {
       res.app.locals['saveDefaultsEnabled'] = enabled
 
       logger.info(`Init service: defaultFilterValuesService: ${res.app.locals['saveDefaultsEnabled']}`)
     }
+  } else {
+    logger.info('SAVE_DEFAULTS_DEBUG No DPR User ID')
   }
 }